set global exception valid

+++ src/main/java/com/todo/payload/ExceptionResponse.java
@@ -15,7 +15,7 @@ public class ExceptionResponse {
 
     private boolean flag;
 
-    private Long statusCode; 
+    private int statusCode;
